refactor: move exceptions and validators to their appropriate files
