+++ sefaria/model/abstract.py
@@ -643,71 +643,3 @@ def cascade_delete_to_list(set_class, fk_attr, pk_attr):
             rec.save()
 
     return foo
-
-class SchemaValidationException(Exception):
-    def __init__(self, key, expected_type):
-        self.key = key
-        self.expected_type = expected_type
-        self.message = f"Invalid value for key '{key}'. Expected type: {expected_type}"
-        super().__init__(self.message)
-class SchemaRequiredFieldException(Exception):
-    def __init__(self, key):
-        self.key = key
-        self.message = f"Required field '{key}' is missing."
-        super().__init__(self.message)
-
-class SchemaInvalidKeyException(Exception):
-    def __init__(self, key):
-        self.key = key
-        self.message = f"Invalid key '{key}' found in data dictionary."
-        super().__init__(self.message)
-
-
-def validate_dictionary(data, schema):
-    """
-    Validates that a given dictionary complies with the provided schema.
-
-    Args:
-        data (dict): The dictionary to be validated.
-        schema (dict): The schema dictionary specifying the expected structure.
-
-    Raises:
-        SchemaValidationException: If the data does not comply with the schema.
-
-    Returns:
-        bool: True if the data complies with the schema, False otherwise.
-    """
-
-    for key, value_type in schema.items():
-        if not (isinstance(value_type, tuple) and len(value_type) == 2 and value_type[1] in ["optional", "required"]):
-            raise ValueError(f"Invalid schema definition for key '{key}'. Use ('type', 'optional') or ('type', 'required').")
-
-    # Check for keys in data that are not in schema
-    for key in data.keys():
-        if key not in schema:
-            raise SchemaInvalidKeyException(key)
-
-    for key, value_type in schema.items():
-        # Check if the key exists in the data dictionary
-        if key not in data:
-            # Check if the field is optional (not required)
-            if isinstance(value_type, tuple) and len(value_type) == 2 and value_type[1] == "optional":
-                continue  # Field is optional, so skip validation
-            else:
-                raise SchemaRequiredFieldException(key)
-
-        # Check if the expected type is a nested dictionary
-        if isinstance(value_type[0], dict):
-            nested_data = data[key]
-            nested_schema = value_type[0]
-            try:
-                # Recursively validate the nested dictionary
-                validate_dictionary(nested_data, nested_schema)
-            except SchemaValidationException as e:
-                # If validation fails for the nested dictionary, re-raise the exception with the key
-                raise SchemaValidationException(f"{key}.{e.key}", e.expected_type)
-
-        # Check the type of the value in the data dictionary
-        elif not isinstance(data[key], value_type[0]):
-            raise SchemaValidationException(key, value_type[0])
-    return True
